Fix memory leak in read_states() by calling pfctl_free_states (#4)
read_states() in pftop.c calls pfctl_get_states() in libpfctl.
That allocates a tail queue (i.e., space for the state entries plus the
overhead for the tail queue itself) and returns that. Then read_states()
runs over that and copies the data into its own structures but it never
cleans up the tail queue.

This commit adds code to free the tail queue after copying the data.

Reported-by: Harold Gutch
Co-authored-by: Oliver Peter <oliver@gfuzz.de>
Co-authored-by: Kristof Provost <kp@FreeBSD.org>